Refactor 'clear_blocks' function
The clear_blocks function was simplified, as there was a good use of recursion
there. However, now there is no early termination when clearing blocks, so
almost all indirect blocks will be explored when clearing. This is not
necessarily bad, but it might be good to add it back in later.