Move most OFT components to level2
All lines that use OFT have been moved to the level2 file, with the exception of
the typedef. It may be better to move the typedef into the file with everything
else, but that will probably be done in a larger effort to refactor global.c and
type.h out into individual files.