Use the __func__ identifier in alloc and dealloc functions
The hard-coded function names in the log statements for the alloc and dealloc
functions has been replaced with the __func__ identifier.

Now there is no need to synchronize the log statement prefixes in these
functions, as is should happen automatically whenever the function name changes.
It is unlikely that the names should ever be changed, but it does not hurt to
make this modification anyhow.