Refactor super and group accessing functions
The super and group accessing functions have been renamed, with the '_block'
suffix removed. The mounting functions have been fixed to access super and group
blocks properly. Some functions have been moved around.

The '_block' suffix was not necessary and added more to type, so I got rid of
it. I forgot to update the mounting functions when the super and group methods
were modified, but now that has been corrected.